fix some default weights

+++ frame/babe/src/default_weights.rs
@@ -26,11 +26,11 @@ impl crate::WeightInfo for () {
 	fn report_equivocation(validator_count: u32) -> Weight {
 		// we take the validator set count from the membership proof to
 		// calculate the weight but we set a floor of 100 validators.
-		let validator_count = validator_count.max(100) as u64;
+		let validator_count = validator_count.max(100) as u32;
 
 		// worst case we are considering is that the given offender
 		// is backed by 200 nominators
-		const MAX_NOMINATORS: u64 = 200;
+		const MAX_NOMINATORS: u32 = 200;
 
 		// checking membership proof
 		(35 * WEIGHT_PER_MICROS)
