+++ frame/staking/src/lib.rs
@@ -3238,8 +3238,8 @@ for Module<T> where
 			});
 
 			if let Some(mut unapplied) = unapplied {
-				let nominators_len = unapplied.others.len() as u64;
-				let reporters_len = details.reporters.len() as u64;
+				let nominators_len = unapplied.others.len() as u32;
+				let reporters_len = details.reporters.len() as u32;
 
 				{
 					let upper_bound = 1 /* Validator/NominatorSlashInEra */ + 2 /* fetch_spans */;
